W160e: mdlFile 00329800 168/144 nd129 -> 144/144 MATCH and 00325d60 window cleared (VU kernel); y_runtime 004542c0 248/256 nd148 -> 256/256 MATCH via YVec3f float-tail retyping + absolute data aliases, 00453ed0 nd402 -> 46

+++ src/Event/mt_evtMessage.c
@@ -978,9 +978,8 @@ void FUN_0039f950(int param_1)
 
       *(u32 *)(*(int *)(param_1 + 0xe8) + iVar3 + 8) = uVar2;
 
-      *(u32 *)(*(int *)(param_1 + 0xe8) + iVar3) =
-
-           *(u32 *)(*(int *)(iVar1 + 0x128) + 0xf4);
+      *(f32 *)(*(int *)(param_1 + 0xe8) + iVar3) =
+           *(f32 *)(*(int *)(iVar1 + 0x128) + 0xf4);
 
       *(int *)(*(int *)(param_1 + 0xe8) + iVar3 + 0xc) = iVar1;
 
@@ -998,9 +997,8 @@ void FUN_0039f950(int param_1)
 
       *(u32 *)(*(int *)(param_1 + 0xe8) + iVar3 + 8) = uVar2;
 
-      *(u32 *)(*(int *)(param_1 + 0xe8) + iVar3) =
-
-           *(u32 *)(*(int *)(iVar1 + 0x128) + 0xf4);
+      *(f32 *)(*(int *)(param_1 + 0xe8) + iVar3) =
+           *(f32 *)(*(int *)(iVar1 + 0x128) + 0xf4);
 
       *(int *)(*(int *)(param_1 + 0xe8) + iVar3 + 0xc) = iVar1;
